Minor cleanup of SwarmLauncher
